Always run all commands in all files.
By running all commands in all files we mimic how circleci works and
you know which are passing and which are failing.

So process-matrix will process all files and prints SUCCESS and exits with
0 if all files pass and prints FAILURE and exits with 1 if any file fails.
It invokes circleci-matrix-notemp.sh.

circleci-matrix-all-commands.sh process all commands and exits with 0 if
all succeed and exits with 1 if any fails.

TODO: Update README.md